# packages/clients/openapi.json
{
  "components": {
    "schemas": {
...
      "CostRowOut": {
        "additionalProperties": false,
        "description": "One grouped cost aggregate: dual metrics never merged (FR-COST-012).",
        "properties": {
          "actual_spend_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Actual Spend Usd",
            "type": "string"
          },
          "cost_estimated_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Estimated Usd",
            "type": "string"
          },
          "cost_partial_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Partial Usd",
            "type": "string"
          },
          "cost_priced_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Priced Usd",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "pricing_version": {
            "title": "Pricing Version",
            "type": "string"
          },
          "subscription_value_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Subscription Value Usd",
            "type": "string"
          },
          "unpriced_event_count": {
            "title": "Unpriced Event Count",
            "type": "integer"
          }
        },
        "required": [
          "key",
          "actual_spend_usd",
          "subscription_value_usd",
          "cost_priced_usd",
          "cost_partial_usd",
          "cost_estimated_usd",
          "unpriced_event_count",
          "pricing_version"
        ],
        "title": "CostRowOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "QueryWarningOut": {
        "additionalProperties": false,
        "description": "A data-quality caveat attached to a v2 query response (FR-QUERY-010).",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "detail",
          "count"
        ],
        "title": "QueryWarningOut",
        "type": "object"
      },
...
      "ReconciliationResponse": {
        "additionalProperties": false,
        "description": "Cost reconciliation drift rows.",
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ReconciliationRowOut"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows"
        ],
        "title": "ReconciliationResponse",
        "type": "object"
      },
      "ReconciliationRowOut": {
        "additionalProperties": false,
        "description": "Observed-versus-computed cost drift for a provider (and optional day).",
        "properties": {
          "computed_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Computed Usd",
            "type": "string"
          },
          "day": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          },
          "drift_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drift Pct"
          },
          "drift_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Drift Usd",
            "type": "string"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "observed_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Observed Usd",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "computed_usd",
          "observed_usd",
          "drift_usd",
          "event_count"
        ],
        "title": "ReconciliationRowOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "tokemetry_server__api__schemas_query__CostResponse": {
        "description": "Cost over a range with the subscription value multiple.",
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          },
          "subscription_monthly_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Monthly Usd"
          },
          "total_cost_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Cost Usd",
            "type": "string"
          },
          "value_multiple": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Multiple"
          }
        },
        "required": [
          "start",
          "end",
          "total_cost_usd",
          "subscription_monthly_usd",
          "value_multiple"
        ],
        "title": "CostResponse",
        "type": "object"
      },
...
      "tokemetry_server__api__v2__schemas__CostResponse": {
        "additionalProperties": false,
        "description": "Grouped costs with data-quality warnings (FR-QUERY-006).",
        "properties": {
          "group_by": {
            "title": "Group By",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CostRowOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/QueryWarningOut"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "group_by",
          "rows",
          "warnings"
        ],
        "title": "CostResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Two API versions coexist during the provider-neutral migration (FR-INGEST-011/014). **v1** (`/api/v1/*`) is the stable collector surface and stays wire-compatible for the whole program. **v2** (`/api/v2/*`) is the provider-neutral surface: batch ingest (`/ingest/events|limits|aggregates`), pre-flight `validate`, the published usage-event JSON schema (`/schemas/usage-event`), and read-only registries. Version is selected by URL path (there is no content negotiation); clients pin a major version and a bearer token, and every response carries an `X-Request-ID`.",
    "summary": "Self-hosted multi-machine AI token usage tracking",
    "title": "tokemetry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cost": {
      "get": {
        "description": "Return total cost over a range and the subscription value multiple.",
        "operationId": "cost_api_v1_cost_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tokemetry_server__api__schemas_query__CostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cost",
        "tags": [
          "query"
        ]
      }
    },
...
    "/api/v2/costs": {
      "get": {
        "description": "Grouped costs with dual metrics, status split, and pricing version.",
        "operationId": "costs_endpoint_api_v2_costs_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "default": "provider",
              "title": "Group By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "machine",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "in": "query",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "unknown_provider",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unknown Provider",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "unknown_model",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unknown Model",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tokemetry_server__api__v2__schemas__CostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Costs Endpoint",
        "tags": [
          "costs"
        ]
      }
    },
    "/api/v2/costs/reconciliation": {
      "get": {
        "description": "Observed-versus-computed cost drift by provider, optionally by day.",
        "operationId": "reconciliation_endpoint_api_v2_costs_reconciliation_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "default": "provider",
              "title": "Group By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "machine",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "in": "query",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "in": "query",
            "name": "unknown_provider",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unknown Provider",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "unknown_model",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unknown Model",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reconciliation Endpoint",
        "tags": [
          "costs"
        ]
      }
    },
...
  }
}